reactored acceptance testing

+++ packages/builder/src/test-request-handler.ts
@@ -1,8 +1,8 @@
 import { Handler, Context } from "./request-handler";
-import { isURL, baseName } from "./path";
+import { isURL } from "./path";
 
-export const testOrigin = "http://localhost:8080";
 const worker = (self as unknown) as ServiceWorkerGlobalScope;
+let testOrigins: string[] = [];
 
 export interface Scenario {
   [fileURL: string]: string;
@@ -15,29 +15,43 @@ export const handleTestRequest: Handler = async function (
   let requestUrl = new URL(req.url);
   let path = requestUrl.pathname;
   let { fs } = context;
+
+  if (testOrigins.includes(requestUrl.origin)) {
+    context.handleOrigin = true;
+    return;
+  }
+
   if (
-    (self as any).testMode &&
-    requestUrl.origin === worker.origin &&
-    // To prevent name collisions between actual JS files being served and JS
-    // files in our testing scenarios that we want to build, use a `test-*.*` for
-    // the name of the test JS files in the testing scenarios
-    !/^test-/.test(baseName(path))
+    path === "/setup-fs" &&
+    req.method === "POST" &&
+    requestUrl.origin === worker.origin
   ) {
-    if (path === "/setup-fs" && req.method === "POST") {
-      await fs.removeAll();
-      let scenario: Scenario = (await req.json()) || {};
-      for (let [path, text] of Object.entries(scenario)) {
-        let file = isURL(path)
-          ? await fs.open(new URL(path), "file")
-          : await fs.open(new URL(path, testOrigin), "file");
-        file.write(text);
-      }
-      return new Response("OK", { status: 200 });
-    } else if (path === "/teardown-fs" && req.method === "POST") {
-      await fs.removeAll();
-      return new Response("OK", { status: 200 });
+    let origin = requestUrl.searchParams.get("origin");
+    if (!origin) {
+      throw new Error(
+        `An origin query parameter must be specified for the test scenario setup`
+      );
     } else {
-      return new Response((await fetch(req)).body);
+      testOrigins.push(origin);
     }
+    await fs.removeAll();
+    let scenario: Scenario = (await req.json()) || {};
+    for (let [path, text] of Object.entries(scenario)) {
+      let file = isURL(path)
+        ? await fs.open(new URL(path), "file")
+        : await fs.open(new URL(path, origin), "file");
+      file.write(text);
+    }
+    return new Response("OK", { status: 200 });
+  } else if (
+    path === "/teardown-fs" &&
+    req.method === "POST" &&
+    requestUrl.origin === worker.origin
+  ) {
+    await fs.removeAll();
+    testOrigins = [];
+    return new Response("OK", { status: 200 });
+  } else if (context.testMode && requestUrl.origin === worker.origin) {
+    return new Response((await fetch(req)).body);
   }
 };
